QueueManager: extract fetch-and-deliver pipeline into FetchedMessages
Move buffer state (_toDeliver, _fetchedPositions, _subscriptions, _nextToRemoveIndex,
_thrownException) and the FetchOnce / WaitForMessageOrTimeout / AfterFlush /
TryDispatch / FailAllSubscriptions operations into a new internal FetchedMessages type.
The reserved EffectId constants and the IdempotencyKeys instance move with the
pipeline. QueueManager keeps lifecycle (Initialize, Dispose, FetchLoop,
CreateQueueClient), the public FetchMessagesOnce / AfterFlush delegates, and the
Subscribe public API which composes WaitForMessageOrTimeout with the post-match
Effect upserts.

MatchResult now carries the resolved ToRemoveId EffectId directly, so QueueManager no
longer needs to know about the PendingDeletion(int) indexing scheme.
